refactor: Phase 3.6 — move 5 monitoring/chat/llm routers to domain modules (#513)
Move audit, usage, monitor → modules/monitoring/
Move chat → modules/chat/
Move llm → modules/llm/

Import replacements (9 total, all in chat.py and llm.py):
- async_chat_manager → chat_service (modules.chat.service)
- async_chat_share_manager → chat_share_service (modules.chat.service)
- async_bot_instance_manager → bot_instance_service (modules.channels.telegram.service)
- async_whatsapp_instance_manager → whatsapp_instance_service (modules.channels.whatsapp.service)
- async_widget_instance_manager → widget_instance_service (modules.channels.widget.service)
- async_cloud_provider_manager → cloud_provider_service (modules.llm.service)
- async_knowledge_collection_manager → knowledge_collection_service (modules.knowledge.service)
- async_audit_logger → audit_service (modules.monitoring.service)

audit.py, usage.py, monitor.py — pure moves (0 db.integration imports).
Original app/routers/ files become thin facade re-exports.
